src/requirements.txt
qrcode
pandas
pillow